requirements.txt
alembic==1.4.2
arrow==0.15.6
astroid==2.2.5
attrs==19.3.0
Authlib==0.14.3
autopep8==1.5.3
awscli==1.18.97
babel==2.8.0
bcrypt==3.1.7
bleach==3.1.4
botocore==1.17.20
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.9.2
dateparser==0.7.4
docutils==0.15.2
ecdsa==0.15
Flask==1.1.2
Flask-CLI==0.4.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
flask-wtf==0.14.3
future==0.17.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
jwt==1.0.0
lazy-object-proxy==1.4.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
moment==0.10
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.1.0
python-jose-cryptodome==1.3.2
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.13
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.3
times==0.7
toml==0.10.1
typed-ast==1.3.5
tzlocal==2.1
urllib3==1.25.9
virtualenv==16.7.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.3.1
zipp==3.1.0
